requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite
requests
PyYAML